fixed the file i was changing and re-did the solution for the reactive state
