+++ nginx.conf
@@ -26,6 +26,8 @@ http {
         default "";  # Default to disallow all origins
         "http://localhost:8080" "http://localhost:8080";
         "http://127.0.0.1:8080" "http://127.0.0.1:8080";
+        "http://localhost:6100" "http://localhost:6100";
+        "http://127.0.0.1:6100" "http://127.0.0.1:6100";
         "https://sherlock-dev.tamucc.edu/semaphore-api" "https://sherlock-dev.tamucc.edu/semaphore-api";
         "https://sherlock-prod.tamucc.edu/semaphore-api" "https://sherlock-prod.tamucc.edu/semaphore-api";
